move case study components to common
